Replace clipping with overlay surfaces
Many browsers do not implement SVG clipping properly. This patch
provides a workaround that does not use clipping to hide geometry but
rather covers up the geometry with polygon filled with the background
color.

To animate the probe, this patch switches from a clipping technique to
actually animating the probe geometry.

The previous clipping-based approach was much simpler and more elegant,
but this workaround seems to be needed for the sake of cross-browser
compatibility.