Fix sea floor generation and add 'maxDepthTilt' config option
The sea floor generation would sometimes overwrite the index with the
predefined max depth and subsequently generate kind of arbitrary max
depths. In rare cases, this would lead to a very shallow sea floor and
positioning the treasure outside the range reachable by the boats.

Overwriting the index has been fixed and the tilt towards a shallow or
deep max depth can be configured via the 'maxDepthTilt' config option.