Fix rendering error that made it possible to spot the treasure location
The sea floor is covered by two background-colored rectangles located
at the left and right of the treasure. Even though the stroke and fill
colors were identical, the stroke color appeared a little darker making
it possible to spot the treasure location. Removing the outline of the
rectangles does not solve the problem completely because the SVG
renderer has issues when SVG rectangles are placed side by side, but at
non-integer coordinates, such that a thin gap is visible in between.
This issue has been addressed by rounding the coordinates appropiately.